% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cells_by_type.R
\name{sampleCellsByType}
\alias{sampleCellsByType}
\title{Sample Cells by Cell Type with Minimum Representation}
\usage{
sampleCellsByType(cell_metadata, n_total_cells = 5000, min_per_type = 50)
}
\arguments{
\item{cell_metadata}{A data frame containing at least two columns:
\code{cell_id} and \code{cell_type}.}

\item{n_total_cells}{Total number of cells to sample.
Default is \code{5000}.}

\item{min_per_type}{Minimum number of cells to sample from each
\code{cell_type}. Default is \code{50}.}
}
\value{
A data frame of sampled cells with at least
\code{min_per_type} cells per \code{cell_type}.
}
\description{
This function samples a specified total number of cells from
a metadata table, ensuring that each cell type is represented
by at least a minimum number of cells.
}
\examples{
meta <- data.frame(
    cell_id = paste0("cell", 1:10000),
    cell_type = sample(c("T", "B", "Mac"), 10000, replace = TRUE)
)
set.seed(42)
sampled <- sampleCellsByType(meta,
    n_total_cells = 1000,
    min_per_type = 50
)

}
