% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankMap-package.R
\docType{package}
\name{RankMap-package}
\alias{RankMap-package}
\title{RankMap: Rank-based reference mapping for fast and robust cell type annotation in spatial and single-cell transcriptomics}
\description{
RankMap is a fast and scalable tool for reference-based cell type annotation of single-cell and spatial transcriptomics data. It uses ranked gene expression and multinomial regression to achieve robust predictions, even with partial gene coverage. Compatible with Seurat, SingleCellExperiment, and SpatialExperiment objects, RankMap offers flexible preprocessing and significantly faster runtime than tools like SingleR, Azimuth, and RCTD.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jinming-cheng/RankMap}
  \item Report bugs at \url{https://github.com/jinming-cheng/RankMap/issues}
}

}
\author{
\strong{Maintainer}: Jinming Cheng \email{jinming.cheng@outlook.com} (\href{https://orcid.org/0000-0003-3806-4694}{ORCID})

}
\keyword{internal}
